+++ TourTogether/.settings/org.eclipse.core.resources.prefs
@@ -1,3 +1,4 @@
 eclipse.preferences.version=1
+encoding//src/main/webapp/WEB-INF/views/index.jsp=UTF-8
 encoding//src/main/webapp/WEB-INF/views/index1.jsp=UTF-8
 encoding//src/main/webapp/WEB-INF/views/member/profile.jsp=UTF-8
