+++ TourTogether/src/main/java/com/cndy/tt/admin/AdminDao.java
@@ -3,6 +3,8 @@ package com.cndy.tt.admin;
 import java.util.Date;
 import java.util.List;
 
+import com.cndy.tt.member.Member;
+
 public interface AdminDao {
 	List<Member> initDateList();
 	List<Integer> initNumList();
