+++ TourTogether/src/main/java/com/cndy/tt/paging/PagingDao.java
@@ -2,16 +2,13 @@ package com.cndy.tt.paging;
 
 import java.util.List;
 
-import com.cndy.tt.admin.Member;
+import com.cndy.tt.member.Member;
 import com.cndy.tt.diary.Diary;
 
 public interface PagingDao {
 
 	public List<Diary> selectPaging(PagingVo pagingVo);
 	public int selectTotalPaging();
-	
-	public boolean insert(Diary diary);
-	public boolean deleteAll();
 	
 	public boolean insert(Diary diary);
 	public boolean deleteAll();
