+++ TourTogether/src/main/java/com/cndy/tt/paging/PagingServiceImpl.java
@@ -5,7 +5,7 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.cndy.tt.admin.Member;
+import com.cndy.tt.member.Member;
 import com.cndy.tt.diary.Diary;
